requirements.txt
appium-python-client
opencv-python
numpy
Pillow